`WorkflowViewStubLifecycleTest` is green.
…and is exercising both `VisualFactory` implementations and a `VisualFactory` based wrapper of `ScreenViewFactoryFinder` for backward compatibility.

Still a lot of TODOs in `AndroidViewFactoryKey` in particular, but that's a massive milestone. Gives me a lot of confidence in this approach.

Biggest bit of remaining uncertainty is all the nooks and crannies of `OverlayDialogHolder`, in support of

- back button hooks
- bounds management
- modal behavior

Also still need to demonstrate Compose integration, but the prototype @helios175 built gives me a lot of confidence there.