Introduces `Wrapper`, `WithName`
First use of `VisualFactory`'s ability to create general purpose factories that are agnostic about specific view and context types. And looky look, I think I've finally found a way to do that without sacrificing support for marker types like `Screen` and `Overlay`.

We introduce this class hierarchy:

```
abstract class Wrapper<W>
  abstract class WithName<W> : Wrapper<W>
    abstract class Named<W> : WithName<W>
    abstract class NamedScreen<W> : WithName<W>, Screen
```

Along with a new `VisualFactory.withName()` extension, a single function that can generate `VisualFactory` functions for simple subtypes of `WithName`.

Also moves `VisualHolder` out to its own file, otherwise unchanged.

`Wrapper` is justified mainly as a convenience to ensure that wrapper renderings include the support for `Compatible` that we almost always forget about. The standardized structure it will invite could also be a boon to test code.

`WithName<W> : Wrapper<W>` is base class for things like `NamedScreen`, and the forthcoming `NamedOverlay` peer that I already need to create for `Dialog`. Would have preferred to call it `Named`, but that name is taken for now.

`VisualFactory.withName()` provides general support for all concrete implementations of `WithName`. It can tranform a factory for renderings of type `R` into one for a particular subtype of `Wrapper<R>`. Note that we're able to delete `VisualFactory.named()` and `VisualFactory.namedScreen()` functions in this PR.

Still need to:
 - Build the same for `EnvironmentScreen`
 - Introduce `NamedOverlay` and `EnvironmentOverlay`.
 - Figure out exactly where these functions should live in the `MultiRendering` hierarchy, taking care that they can be reconfigured

`WorkflowViewStubLifecycleTest` is still green. It uses `WorkflowUiTestActivity.setRendering`, which relies on `NamedScreen`, so yes we're sure the new code is being exercised.

A little more context on why this is worth doing. It wouldn't need to be anywhere near so complicated if we were willing to abandon our `Screen` and `Overlay` marker interfaces, e.g. change `WorkflowViewStub.show(Screen)` back to `WorkflowViewStub.show(Any)`. But that compile-time distinction between view-like-things (`Screen`) and `window-like-things` (`Overlay`) is pretty valuable when building features, and it would be a shame to lose it -- especially if we are able to keep it from leaking into the fundamental machinery like `AndroidViewFactory` and `AndroidViewMultiRendering`, used mainly under the hood. What's exciting about the approach in this PR is that with a little bit of ceremony we get to keep the marker types without creating entire parallel hierarchies of copy / paste factory implementations.