auto-update Dependency Guard baselines for Renovate PRs
This adds a new workflow to the `kotlin.yml` GitHub Action.

If the PR was made by Renovate, automatically update baselines and push so that no one has to check out the branch and update the baselines manually.

If a non-bot made the pull request, run the non-baseline task which fails on changes.

If dependency-guard generated changes, commit and push those changes. This relies upon the 'cancel-stale-jobs' job to cancel the rest of the jobs for the now-stale commit.