Nested overlays sample uses overlays.
And can demonstrate the bug where dialogs are shown out of order:

- Click _Cover Everything_
- Click _Cover Body_

The red inner dialog is shown over the outer green dialog,
but the green one should always be on top. (#966)

Also introduces `name` parameter for `BodyAndOverlaysScreen`, because they're
a nightmare to nest without it. See the kdoc for details.