prioritize `JavaCompile.Options.release` and Android `CompileOptions` when parsing an artifact jdk version
There are several methods for setting a JDK version, but the artifacts check plugin is only parsing the lowest priority method.  `JavaPluginExtension.targetCompatibility.majorVersion` can be set directly, or via the newer `javaToolchain` functionality.  However, in an Android project, the `compileOptions.targetCompatibility` property has a higher priority, and the `options.release` property in a `JavaCompile` task is higher than them both.

Prior to these changes, if a JDK version setting was overwritten by the other methods, the plugin wouldn't catch it.  Now it will.

I also made the decision to change `ArtifactConfig.javaVersion` from a `String` to an `Int`.  This is because there's ambiguity when the version is a String and it's representing Java 8.  In this case, is the string version just `"8"` or is it `"1.8"`?  The different methods report different versions.  Just treating the version as an `Int` removes that ambiguity.