Introduces `ScreenTransitionLogger`
`ScreenTransitionLogger` provides extremely minimal hooks to allow apps to implement whatever it is that "navigation logging" means to them. It is a `ViewEnvironment` service with a default no-op implementation.

`ScreenTransitionLogger.onTransition` is called whenever one of our stock stub views (`WorkflowViewStub` and `@Composable fun WorkflowRendering()`) is updated to:
- show its first `Screen` rendering, or to
- show one that is incompatible with whatever was last shown

Both the incoming and outgoing `Screen` are provided to `onTransition`. The latter may be null when the function is called with the first `Screen` a stub is showing.

For this first stab, we are very intentionally providing a very unambitious mechanism.
- We make no attempt to report the entire AndroidX lifecycle -- e.g., we do not report when stubs are set up or torn down, only when they show something new.
- We focus only on `Screen`, not `Overlay`. In practice we have found that all of our interesting `Overlay` implementations are based on `ScreenOverlay` and stubs anyway.
- We do not attempt to report that view code is "done" updating, only that we have determined that a new view is required.

Note that this approach _will_ log updates under `BackStackScreen`, because `BackStackContainer` is built around `WorkflowViewStub`. Updates to the UI root will also be reported, because `WorkflowLayout` wraps `WorkflowViewStub` as well.

It is worth noting what _is_ supported: `onTransition` is provided with the `ViewEnvironment` in addition to the incoming and outgoing `Screen`s. This should allow more complex apps to set up logging systems that reflect more complex structures. For example, a custom container tiling or layering multiple stubs could provide `ViewEnvironment` values naming these separate navigation contexts. Or if a particular custom container is very fine grained in its use of stubs, it can seed its children's environments with a no-op `ScreenTransitionLogger` to eliminate noise.