adding custom actions for CI
This adds some custom internal actions, which removes some boilerplate from job definitions and ensures that everything's consistent.

For each step which invokes a Gradle task, we will:

- set up JDK 11 (zulu)
- set Gradle properties and jvm args, where the memory settings are OS-specific
  - `-Xmx10g` for MacOS
  - `-Xmx4g` for Ubuntu
  - `-Xmx4g` for Windows
- use `gradle-build-action@v2` for caching
- invoke whatever Gradle task(s)
- upload a heap dump in the event of a memory error

This also adds a `gradle-task-with-commit` action which does all the above, except it will automagically run format/baseline tasks before a "check" task if it has permissions to push to the PR branch.  So it will automatically fix KtLint issues.