use the `build-logic` and `main-build-artifacts` caches for snapshots
GitHub Actions caches have scopes.  For any given pull request, it can download its own caches and it can download the caches associated with the `main` branch.  Once that PR branch is merged, its cache is useless.

By using and writing to these two caches on the `main` branch, we speed up the `publish` builds but we also keep the caches as up-to-date as possible.  In turn, that means more build cache hits and faster builds for all pull requests.