Cache the AVD in CI before executing any tests
background: https://workflow-community.slack.com/archives/CHTFPR277/p1689193974005269

I was using `actions/cache@v3` to restore and possibly save the AVD in instrumented test shards.

That basic "cache" action handles both reading and writing to the cache.  The writing is done during a post-action, meaning that if the AVD was cached, it was cached *after* it had been used for the instrumented tests.

Now, we use the newer `actions/cache/restore@v3` and `actions/cache/save@v3` actions, like we use elsewhere.  Now, if there was a cache miss and we have to make a new AVD, it is cached before the tests are executed so that it doesn't have any of our test apps installed.