Fixes leak when calling `destroyOnDetach` from `onDetachedFromWindow`.
`View.isAttachedToWindow` returns false while the view is dispatching calls to `onDetachedFromWindow`. As a result, calling `WorkflowLifecycleOwner.destroyOnDetach` from an `onDetachedFromWindow` handler that was downstream of its own would mean that the WLO would never get destroyed.

Also makes `DialogSession.destroyDialog` idempotent. Necessary because `ComponentDialog.dismiss()` is not, and with this fix we are more likely to call that redundantly.

(The actual culprit is `ComponentDialog.onStop()`, which nulls out `ComponentDialog._lifecycleRegistry`. The second call to that method creates a new instance on demand as a side effect to its call to `handleLifecycleEvent`, which immediately throws due to an illegal transition from `INITIALIZED` to `DESTROYED`. Tried to rely on just checking `Dialog.isShowing` before calling `dismiss()`, but that's not reliable.)