1247: Partial Tree Rerendering
If PARTIAL_TREE_RENDERING is in the RuntimeConfig then,
Track whether or not the state (or state of a descendant) changed in the WorkflowNode. Pass lastRendering if its not.

Also adds tests for this behavior and expands the test runtime matrix.
Also adds shards for the new runtime for instrumentation tests.