Introduces Activity.workflowContentView, workflowContentViewOrNull
Simplifies creating a `WorkflowLayout` and using it as the `contentView` of an `Activity`. Written with an eye toward making it easy to hand off from one workflow runtime to another, or doing something like `activity.workflowContentView.show(LoadingScreen)` before any workflow runtime is put in place. Nice for testing too.