Edit composables to align with their responsibilities
The order of the main content being displayed was from parsing -> rendering, which would lead to complexity when adding in tabbing functionality.
Having the order be Renderer -> parsing -> rendering was more clear.
This commit also introduces the idea of a "trace", where we have input a json array instead of a singular json