[DNM] Experimenting with a state queue idea.
I thought of a interesting idea for the emitOutput issue: Build a snapshot state queue and make emitOutput just write to the queue, so it behaves like any other compose state write and gets sent to the workflow runtime at the same time as other state changes. When the compose workflow node gets the frame request from the state change, it can check dequeue from the output queue and send the output up the action cascade. If multiple things are queued, we can just check the queues again when we're building up the SelectBuilder and if there's something ready to be processed create a one-off CompletableDeferred with it to complete the select immediately.

this would only work if we ensured a certain ordering of callbacks that might be a brittle thing to rely on, and there is other overhead we'd need so might not have any perf impact. Only potential benefit would be that outputs and state changes would dispatch to the workflow runtime through the same mechanism. I've basically convinced myself the extra complexity to implement isn't worth it.