Make RenderingProduced not a value class.
Value classes get boxed as soon as they're used as a supertype, so it's basically pointless to make sealed class children value classes. I made it a data class instead.

This change is required to move `:workflow-runtime-android` into `:workflow-runtime`, or we get a compiler crash.

I also made the object siblings data objects to fix a compiler suggestion and get better `toString` methods.