Move SnapshotParcels.kt to workflow-core.
This gives workflow-core an android source set and moves the Parcelable-related
Snapshot helpers to it. This makes more sense since the other Snapshot helpers
are in the core module, and Parcelable isn't a UI thing it's just an android thing.

This also paves the way for go/compose-based-workflow to use these helpers to
support `rememberSaveable`.